completed section-4, implemented fluent validation lib

+++ HR.LeaveManagement.Application/Features/LeaveType/Command/DeleteLeaveType/DeleteLeaveTypeCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using HR.LeaveManagement.Application.Contracts.Persistance;
+using HR.LeaveManagement.Application.Exceptions;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -15,6 +16,12 @@ namespace HR.LeaveManagement.Application.Features.LeaveType.Command.DeleteLeaveT
         {
             //retrive the record from the database
             var leaveTypeToDelete = await leaveTypeRepository.GetByIdAsync(request.Id);
+
+            //verify the record exists
+            if (leaveTypeToDelete == null)
+            {
+                throw new NotFoundException(nameof(LeaveType), request.Id);
+            }
 
             //remove from the database.
            await leaveTypeRepository.DeleteAsync(leaveTypeToDelete);
